chore: Update imports in page.tsx to reflect renamed signup.tsx
